requirements.txt
kubernetes
tornado
pyjwt
pyyaml
jinja2
mysql-connector-python-rf
